requirements.txt
backports.ssl-match-hostname==3.4.0.2
basemap==1.0.7
certifi==14.5.14
Cython==0.22
emcee==2.1.0
Flask==0.10.1
Flask-SocketIO==0.6.0
flickrapi==2.0
folium==0.1.4
gensim==0.10.3
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
gnureadline==6.3.3
greenlet==0.4.5
h5py==2.5.0
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
jsonschema==2.4.0
MarkupSafe==0.23
matplotlib==1.3.1
microsofttranslator==0.7
mistune==0.5.1
mock==1.0.1
MySQL-python==1.2.5
nltk==3.0.1
nose==1.3.3
numexpr==2.4
numpy==1.9.2
oauthlib==0.7.2
pandas==0.15.2
patsy==0.3.0
Pillow==2.8.1
protobuf==2.6.1
ptyprocess==0.5
Pygments==2.0.2
pymongo==2.8
pymstranslator==0.0.4
PyMySQL==0.6.6
pyparsing==2.0.2
python-dateutil==2.4.1
pytz==2014.10
PyYAML==3.11
pyzmq==14.6.0
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
requests-toolbelt==0.4.0
scikit-image==0.10.1
scikit-learn==0.15.2
scipy==0.15.1
seaborn==0.5.1
setproctitle==1.1.8
six==1.9.0
SQLAlchemy==1.0.5
statsmodels==0.5.0
stevedore==0.15
sympy==0.7.6
terminado==0.5
Theano==0.6.0
tornado==4.1
tweepy==3.3.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
Werkzeug==0.10.1
